# Optimizing Brilirs print