--- a/benchmarks/.gitignore
+++ b/benchmarks/.gitignore
@@ -1,3 +1,3 @@
-*.bench.json
+**/*.bench.json
 bench.csv
 plot.svg