--- a/benchmarks/Makefile
+++ b/benchmarks/Makefile
@@ -1,13 +1,16 @@
-BENCHMARKS := $(wildcard *.bril)
+BENCHMARKS := core/*.bril \
+			float/*.bril \
+			mem/*.bril \
+			mixed/*.bril
 
 .PHONY: bench clean plot
 bench:
 	turnt -e bench --save $(BENCHMARKS)
 clean:
-	rm -f *.bench.json plot.svg bench.csv
+	rm -f **/*.bench.json plot.svg bench.csv
 plot: plot.svg
 
-bench.csv: $(wildcard *.bench.json)
+bench.csv: $(wildcard **/*.bench.json)
 	python3 summarize.py $^ > $@
 
 %.svg: %.vl.json bench.csv