--- a/benchmarks/function_call.prof
+++ b/benchmarks/function_call.prof
@@ -1,1 +1,0 @@
-total_dyn_inst: 48607906