--- a/benchmarks/turnt.toml
+++ b/benchmarks/turnt.toml
@@ -22,7 +22,9 @@ command = "bril2json < {filename} | ../brilift/target/release/brilift -j -- {arg
 default = false
 command = """
 bril2json < {filename} > {base}.json
-make -C ../brilift ../benchmarks/{base}
+NAME=$(dirname {filename})/{base}
+make -C ../brilift ../benchmarks/$NAME
+mv $NAME {base}
 
 hyperfine --warmup 3 --export-json bench.json \
 'brili {args} < {base}.json' \
@@ -30,5 +32,6 @@ hyperfine --warmup 3 --export-json bench.json \
 '../brilirs/target/release/brilirs {args} < {base}.json' \
 './{base} {args}'
 
-rm -f {base}.json"""
-output."bench.json" = "bench.json"
+rm -f {base}.json
+rm {base}"""
+output."bench.json" = "../bench.json"