--- a/brilift/Makefile
+++ b/brilift/Makefile
@@ -5,8 +5,8 @@
 
 # Brilift only supports core Bril for now, so we select those tests &
 # benchmarks.
-TESTS := ../test/interp/core/*.bril ../test/interp/float/*.bril ../test/interp/mem/*.bril
-BENCHMARKS := ../benchmarks/*.bril
+TESTS := ../test/interp/core/*.bril ../test/interp/float/*.bril ../test/interp/mem/*.bril ../test/interp/mixed/*.bril
+BENCHMARKS := ../benchmarks/core/*.bril ../benchmarks/float/*.bril ../benchmarks/mem/*.bril ../benchmarks/mixed/*.bril
 
 CFLAGS := $(if $(TARGET),-target $(TARGET))
 BRILFLAGS := $(if $(TARGET),-t $(TARGET))