--- a/brilirs/Cargo.toml
+++ b/brilirs/Cargo.toml
@@ -21,6 +21,7 @@ thiserror    = "1.0"
 clap         = { version = "3.2", features = ["derive"] }
 fxhash       = "0.2"
 mimalloc     = "0.1"
+itoa         = "1.0"
 
 [dependencies.bril-rs]
 version      = "0.1.0"