--- a/brilirs/src/interp.rs
+++ b/brilirs/src/interp.rs
@@ -157,21 +157,14 @@ impl Heap {
   }
 }
 
-// A getter function for when you just want the Value enum
-fn get_value<'a>(vars: &'a Environment, index: usize, args: &[usize]) -> &'a Value {
-  vars.get(&args[index])
-}
-
 // A getter function for when you know what constructor of the Value enum you have and
 // you just want the underlying value(like a f64).
-fn get_arg<'a, T>(vars: &'a Environment, index: usize, args: &[usize]) -> T
-where
-  T: From<&'a Value>,
-{
+// Or can just be used to get a owned version of the Value
+fn get_arg<'a, T: From<&'a Value>>(vars: &'a Environment, index: usize, args: &[usize]) -> T {
   T::from(vars.get(&args[index]))
 }
 
-#[derive(Debug, Default, Clone)]
+#[derive(Debug, Default, Clone, Copy)]
 enum Value {
   Int(i64),
   Bool(bool),
@@ -181,7 +174,7 @@ enum Value {
   Uninitialized,
 }
 
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq, Copy)]
 struct Pointer {
   base: usize,
   offset: i64,
@@ -212,11 +205,7 @@ impl fmt::Display for Value {
 
 fn optimized_val_output<T: std::io::Write>(out: &mut T, val: &Value) -> Result<(), std::io::Error> {
   match val {
-    Value::Int(i) => {
-      let mut buf = itoa::Buffer::new();
-      let s = buf.format(*i);
-      out.write_all(s.as_bytes())
-    }
+    Value::Int(i) => out.write_all(itoa::Buffer::new().format(*i).as_bytes()),
     Value::Bool(b) => out.write_all(b.to_string().as_bytes()),
     Value::Float(f) if f.is_infinite() && f.is_sign_positive() => out.write_all(b"Infinity"),
     Value::Float(f) if f.is_infinite() && f.is_sign_negative() => out.write_all(b"-Infinity"),
@@ -224,9 +213,7 @@ fn optimized_val_output<T: std::io::Write>(out: &mut T, val: &Value) -> Result<(
     Value::Float(f) => out.write_all(format!("{f:.17}").as_bytes()),
     Value::Pointer(p) => out.write_all(format!("{p:?}").as_bytes()),
     Value::Uninitialized => unreachable!(),
-  }?;
-  // Add new line
-  out.write_all(&[b'\n'])
+  }
 }
 
 impl From<&bril_rs::Literal> for Value {
@@ -289,6 +276,12 @@ impl<'a> From<&'a Value> for &'a Pointer {
   }
 }
 
+impl From<&Self> for Value {
+  fn from(value: &Self) -> Self {
+    *value
+  }
+}
+
 // Sets up the Environment for the next function call with the supplied arguments
 fn make_func_args<'a>(callee_func: &'a BBFunction, args: &[usize], vars: &mut Environment) {
   vars.push_frame(callee_func.num_of_vars);
@@ -297,8 +290,8 @@ fn make_func_args<'a>(callee_func: &'a BBFunction, args: &[usize], vars: &mut En
     .iter()
     .zip(callee_func.args_as_nums.iter())
     .for_each(|(arg_name, expected_arg)| {
-      let arg = vars.get_from_last_frame(arg_name).clone();
-      vars.set(*expected_arg, arg);
+      let arg = vars.get_from_last_frame(arg_name);
+      vars.set(*expected_arg, *arg);
     });
 }
 
@@ -379,7 +372,7 @@ fn execute_value_op<'a, T: std::io::Write>(
       state.env.set(dest, Value::Bool(arg0 || arg1));
     }
     Id => {
-      let src = get_value(&state.env, 0, args).clone();
+      let src = get_arg::<Value>(&state.env, 0, args);
       state.env.set(dest, src);
     }
     Fadd => {
@@ -445,8 +438,7 @@ fn execute_value_op<'a, T: std::io::Write>(
           .iter()
           .position(|l| l == last_label)
           .ok_or_else(|| InterpError::PhiMissingLabel(last_label.to_string()))
-          .map(|i| get_value(&state.env, i, args))?
-          .clone();
+          .map(|i| get_arg::<Value>(&state.env, i, args))?;
         state.env.set(dest, arg);
       }
     },
@@ -458,7 +450,7 @@ fn execute_value_op<'a, T: std::io::Write>(
     Load => {
       let arg0 = get_arg::<&Pointer>(&state.env, 0, args);
       let res = state.heap.read(arg0)?;
-      state.env.set(dest, res.clone());
+      state.env.set(dest, *res);
     }
     PtrAdd => {
       let arg0 = get_arg::<&Pointer>(&state.env, 0, args);
@@ -472,13 +464,14 @@ fn execute_value_op<'a, T: std::io::Write>(
 
 fn execute_effect_op<'a, T: std::io::Write>(
   state: &'a mut State<T>,
-  func: &BBFunction,
   op: bril_rs::EffectOps,
   args: &[usize],
   funcs: &[usize],
   curr_block: &BasicBlock,
+  // There are two output variables where values are stored to effect the loop execution.
   next_block_idx: &mut Option<usize>,
-) -> Result<Option<Value>, InterpError> {
+  result: &mut Option<Value>,
+) -> Result<(), InterpError> {
   use bril_rs::EffectOps::{
     Branch, Call, Commit, Free, Guard, Jump, Nop, Print, Return, Speculate, Store,
   };
@@ -492,19 +485,17 @@ fn execute_effect_op<'a, T: std::io::Write>(
       *next_block_idx = Some(curr_block.exit[exit_idx]);
     }
     Return => {
-      return Ok(
-        func
-          .return_type
-          .as_ref()
-          .map(|_| get_value(&state.env, 0, args).clone()),
-      )
+      if !args.is_empty() {
+        *result = Some(get_arg::<Value>(&state.env, 0, args));
+      }
     }
     Print => {
       // In the typical case, users only print out one value at a time
       // So we can usually avoid extra allocations by providing that string directly
       if args.len() == 1 {
-        optimized_val_output(&mut state.out, state.env.get(args.get(0).unwrap()))
-          .map_err(|e| InterpError::IoError(Box::new(e)))?;
+        optimized_val_output(&mut state.out, state.env.get(args.get(0).unwrap()))?;
+        // Add new line
+        state.out.write_all(&[b'\n'])?;
       } else {
         writeln!(
           state.out,
@@ -514,8 +505,7 @@ fn execute_effect_op<'a, T: std::io::Write>(
             .map(|a| state.env.get(a).to_string())
             .collect::<Vec<String>>()
             .join(" ")
-        )
-        .map_err(|e| InterpError::IoError(Box::new(e)))?;
+        )?;
       }
     }
     Nop => {}
@@ -529,8 +519,8 @@ fn execute_effect_op<'a, T: std::io::Write>(
     }
     Store => {
       let arg0 = get_arg::<&Pointer>(&state.env, 0, args);
-      let arg1 = get_value(&state.env, 1, args);
-      state.heap.write(arg0, arg1.clone())?;
+      let arg1 = get_arg::<Value>(&state.env, 1, args);
+      state.heap.write(arg0, arg1)?;
     }
     Free => {
       let arg0 = get_arg::<&Pointer>(&state.env, 0, args);
@@ -538,7 +528,7 @@ fn execute_effect_op<'a, T: std::io::Write>(
     }
     Speculate | Commit | Guard => unimplemented!(),
   }
-  Ok(None)
+  Ok(())
 }
 
 fn execute<'a, T: std::io::Write>(
@@ -548,6 +538,7 @@ fn execute<'a, T: std::io::Write>(
   let mut last_label;
   let mut current_label = None;
   let mut curr_block_idx = 0;
+  // A possible return value
   let mut result = None;
 
   loop {
@@ -559,12 +550,8 @@ fn execute<'a, T: std::io::Write>(
     last_label = current_label;
     current_label = curr_block.label.as_ref();
 
-    // This helps to implement fallthrough with basic blocks when there is no control flow instruction at the end of the block
-    let mut next_block_idx = if curr_block.exit.len() == 1 {
-      Some(curr_block.exit[0])
-    } else {
-      None
-    };
+    // A place to store the next block that will be jumped to if specified by an instruction
+    let mut next_block_idx = None;
 
     for (code, numified_code) in curr_instrs.iter().zip(curr_numified_instrs.iter()) {
       match code {
@@ -622,21 +609,25 @@ fn execute<'a, T: std::io::Write>(
           funcs: _,
           pos,
         } => {
-          result = execute_effect_op(
+          execute_effect_op(
             state,
-            func,
             *op,
             &numified_code.args,
             &numified_code.funcs,
             curr_block,
             &mut next_block_idx,
+            &mut result,
           )
           .map_err(|e| e.add_pos(*pos))?;
         }
       }
     }
+
+    // Are we jumping to a new block or are we done?
     if let Some(idx) = next_block_idx {
       curr_block_idx = idx;
+    } else if curr_block.exit.len() == 1 {
+      curr_block_idx = curr_block.exit[0];
     } else {
       return Ok(result);
     }
@@ -758,17 +749,14 @@ pub fn execute_main<T: std::io::Write, U: std::io::Write>(
     return Err(InterpError::MemLeak).map_err(|e| e.add_pos(main_func.pos));
   }
 
-  state
-    .out
-    .flush()
-    .map_err(|e| InterpError::IoError(Box::new(e)))?;
+  state.out.flush().map_err(InterpError::IoError)?;
 
   if profiling {
     writeln!(profiling_out, "total_dyn_inst: {}", state.instruction_count)
       // We call flush here in case `profiling_out` is a https://doc.rust-lang.org/std/io/struct.BufWriter.html
       // Otherwise we would expect this flush to be a nop.
       .and_then(|_| profiling_out.flush())
-      .map_err(|e| InterpError::IoError(Box::new(e)))?;
+      .map_err(InterpError::IoError)?;
   }
 
   Ok(())