… benchmarks/core/ackermann.bril
… benchmarks/core/ackermann.out
… benchmarks/core/ackermann.prof
… benchmarks/core/armstrong.bril
… benchmarks/core/armstrong.out
… benchmarks/core/armstrong.prof
… benchmarks/core/binary-fmt.bril
… benchmarks/core/binary-fmt.out
… benchmarks/core/binary-fmt.prof
… benchmarks/core/bitwise-ops.bril
… benchmarks/core/bitwise-ops.out
… benchmarks/core/bitwise-ops.prof
… benchmarks/core/catalan.bril
… benchmarks/core/catalan.out
… benchmarks/core/catalan.prof
… benchmarks/core/check-primes.bril
… benchmarks/core/check-primes.out
… benchmarks/core/check-primes.prof
… benchmarks/core/collatz.bril
… benchmarks/core/collatz.out
… benchmarks/core/collatz.prof
… benchmarks/core/digital-root.bril
… benchmarks/core/digital-root.out
… benchmarks/core/digital-root.prof
… benchmarks/core/euclid.bril
… benchmarks/core/euclid.out
… benchmarks/core/euclid.prof
… benchmarks/core/factors.bril
… benchmarks/core/factors.out
… benchmarks/core/factors.prof