Working off of #215, I was kind of curious how much easy performance I could get out of my modified `function_call.bril` benchmark. Some of the easy wins are in avoiding a `Vec` allocation for single argument prints, avoiding `std::fmt/writeln!` as much as possible by just writing bytes to `stdout`, using `&str` instead of allocating a `String` when possible, and using https://crates.io/crates/itoa/1.0.2 for converting integers to `&str`.

| `brili` | old `brilirs` | new `brilirs` | `brilift -j` | `brilift` aot |
| -- | ---| --- | ----| --- | 
| 14.289 s ±  0.140 s | 674.1 ms ±  14.4 ms | 416.3 ms ±   8.1 ms | 1.776 s ±  0.026 s | 1.069 s ±  0.006 s |

I've also included the update I've made to `function_call.bril` as the second commit. By having a side-effect with each function call, it makes the benchmarking of interpreters more interesting since the function call can't necessarily be optimized away. However, it also makes an already long running benchmark take even longer and has a large amount of output which makes it great for seeing small interpreter improvements but maybe not ideal to have as a general benchmark(plus the output file is 5.6M lines long/11MB). Maybe this commit should not be included or moved somewhere else.

Disclaimer: This is a silly benchmark which is probably never representative of an actual Bril program.