@Pat-Lafon 2022-06-27
One would similarly like like to use the https://docs.rs/ryu/latest/ryu/index.html crate for converting floats to strings(based on the PLDI paper of the same name). However, the implementation does not allow for specifying the necessary precision of 17 so I gave up trying to use that(though I got a great performance benefit when testing it). In doing so, I found a small bug in brilift and have included the fix + an improved test case.