@Pat-Lafon 2022-07-01
I'm not sure if I messed something up in terms of the benchmarking scripts used to make the plot. I am able create plots from the benchmarking times but the `brilift-aot` time is a lot slower than I would of expected(roughly a factor of 2 slower than `bril-jit`). I'm not sure why, maybe it's just something with the setup on my end, but this part in particular probably needs a second look over.